requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
xarray>=2023.1.0
netCDF4>=1.6.0
scipy>=1.10.0
plotly>=5.15.0

# Geographic/mapping
geopandas>=0.14.0
cartopy>=0.22.0
shapely>=2.0.0

# Large datasets
dask>=2023.1.0

# Plotting
matplotlib>=3.7.0

# Pattern Detection Engine
scikit-learn>=1.3.0
tsfresh>=0.20.2
mlxtend>=0.23.1
pyod>=2.0.0
imbalanced-learn>=0.11.0
statsmodels>=0.14.0

# Causal Discovery
tigramite>=5.2.0
networkx>=3.0
causal-learn>=0.1.3.8  # PC, LiNGAM, NOTEARS algorithms (Issue #9)

# API Backend
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# LLM Integration (Ollama)
ollama>=0.1.0

# Knowledge Graph Databases
neo4j>=5.0.0           # Neo4j driver with vector support
surrealdb>=0.3.0       # SurrealDB Python driver

# Optional: Vector DB for RAG
# chromadb>=0.4.0
# sentence-transformers>=2.2.0

# Legacy UI (optional)
streamlit>=1.28.0
gradio>=4.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0  # for FastAPI TestClient

# Configuration
pydantic-settings>=2.0.0

# Logging
structlog>=23.0.0
python-json-logger>=2.0.0
slowapi>=0.1.9
